[sw] Allow for the optional inclusion of a .bazelrc-site file
The .bazelrc-site file allows for site-specific compile options
(output roots, library paths etc).   This file is an optional
addition to each working repository, however this file should
not be added to the upstrem repository. Therefore this file
is also added to .gitignore, and a CI check is added to
entsure that this file is not forcibly added to git where
it could contaminate the site-specific setup of other users.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>