[chip, dv] Introduce chip_if
This commit introduces chip_if.sv, which will house all
functional interfaces internally. Only the `ios` wires
in chip_if will connect directly to the DUT. All sub-interfaces
in chip_if are made internally or externally gateable to
support multiple functions muxed onto the same IOs.

chip_if needs access to some parameters and types. So
this commit also adds chip_common_pkg.sv, which is imported
in chip_if as well as chip_env_pkg. The changes to other files
support this change.

chip_if is bound to the DUT in tb and passed on to the
UVM testbench. But it is not functional in this commit. In
subsequent commits, the testbench is incrementally updated
to use the chip_if interfaces in lieu of the connections
and hard-wired interfaces in tb.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>