[test] Fix rom_e2e_bootstrap_enabled_not_requested
This commit updates `rom_e2e_bootstrap_enabled_not_requested` (#14451)
to use `READ_SFDP` instead of `READ_STATUS` to verify that the target is
not in bootstrap mode.

CIPO line is in high-Z mode when CMD_INFO registers are not configured.
Thus, we use `READ_STATUS` to avoid false negatives when bootstrap is
not requested.

Signed-off-by: Alphan Ulusoy <alphan@google.com>