[chip, pwm dv] Fix chip level PWM sleep test
While debugging the chip level PWM sleep test, I ended up
refactoring the PWM agent. The changes to the PWM monitor
are cosmetic, towards improved efficiency and reuse.

At the chip level, the key change is to the clock and reset
ports connected to the PWM agent's interface. I accidentally
had connected the bus clock instead of the AON clock, which
resulted in mismatches.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>