[bazel] fix bug in the bitstream cache manager
PR #14938 identified a bug in the Bazel bitstreams cache manager script
(`rules/scripts/bitstreams_workspace.py`). Specifically, if the cache
is not yet initialized, and the default bitstream key "HEAD" is provided,
then no local "latest.txt" file will get created to indicate what is the
latest bitstream in the local cache.

Note, the git commit hash ID of the bitstream in the local `latest.txt`
is the "latest" bitstream currently in the local cache on the host, and
may differ from the "latest" bitstream in the GCS bucket if your
current branch is behind the most recent commit to master, which is why
the bitstream cache manager script defaults to pulling down the "latest"
bitstream behind the current HEAD commit.

This should get the Bazel airgapped images autogenerating successfully
again (the job is currently failing due to this bug).

Signed-off-by: Timothy Trippel <ttrippel@google.com>