[bazel] Add 'build-then' command to bazelisk.sh
This is just a convenience feature that makes it easier to build a
target and then do something with the output files.

I found myself frequently building disassemblies, using "./bazelisk.sh
outquery" to find the filename, and then opening the output file in
"less". This workflow can now be accomplished with a command like this:

    ./bazelisk.sh build-then "less %s" //sw/device/silicon_creator/rom:rom_fpga_cw310_dis

Signed-off-by: Dan McArdle <dmcardle@google.com>