feat(chip_if): Add sleep state indicator.
pwrmgr_low_power_if.low_power is not the exact indicator of the chip low
power state. The signal is asserted when the fast FSM in PWRMGR starts
moving into low power state.

This commit adds `in_sleep` to the interface and the signal is asserted
when the fast FSM moves to `FastPwrStateLowPower`.

Signed-off-by: Eli Kim <eli@opentitan.org>