[rom/e2e] add smoke testpoint and map unmapped test in DV
The simplest of ROM E2E tests, that checked if ROM could successfully
boot and execute a test (that simple returned `true`) via the OTTF using
default test infrastructure configurations, was running in DV but was
not mapped to a test point in the ROM E2E testplan. This adds a
testpoint for said test and fixes #14679.

Signed-off-by: Timothy Trippel <ttrippel@google.com>