[bazel] make go toolchain dependency config work in airgapped env
This updates the way the go toolchain is downloaded/included in our
Bazel workspace to enable Bazel builds in an airgapped environment.
Specifically, #15044 exposed a bug in our go toolchain dependency that
caused Bazel builds to fail in the airgapped environment.

The fix for this was to manually specify the exact toolchain tarballs
to be downloaded using the `go_download_sdk` repo rule, instead of the
`go_register_toolchain` rule which attempts to automatically find the
right toolchain to download by downloading a file from the Internet that
is not cacheable, since no sha256 is provided to the `ctx.download`
action, and therefore not accessible in an airgapped environment.

Signed-off-by: Timothy Trippel <ttrippel@google.com>