[rv_dm] Declare RV_DM debug memory as RW region
Since the RV_DM debug memory contains both writeable and readable regions,
the corresponding window in rv_dm.hjson needs to be marked accordingly.

Note that that window is only a stub and cannot be used to build an accurate
RAL model (we model the memory and registers inside rv_dm more accurately
in rv_dm_debug_mem.hjson for simulation purposes).

Signed-off-by: Michael Schaffner <msf@google.com>