[dv/pwrmgr] Fixes for recent RTL changes
The change in handling of reset had a big impact on DV.
Some test sequences need to be adjusted to stop during lc reset or clock
stops, and reset handling needs to change.
The pwrmgr_clk_ctrl_agent was out of sync with the counterpart in
pwrmgr_base_vseq, so we are moving all functionality to the latter.
This PR disables that agent and a subsequent one will remove it.

Signed-off-by: Guillermo Maturana <maturana@google.com>