[crypto] Safely generate the P-256 secret scalar and random keys.
Previously, the secret scalar k was simply fetched from RND, but when
reduced modulo the curve order the value will be biased. For this
reason, the FIPS spec requires generating the secret scalar from a
larger (320-bit) seed, and additionally ensuring it is nonzero.
Generation of the secret key from a random seed is very similar.

Signed-off-by: Jade Philipoom <jadep@google.com>